refactor: extract admin debug query contract
